refactor: from 메소드 이동
