+++ monicar-collector/src/main/java/org/collector/presentation/dto/CListRequest.java
@@ -2,8 +2,6 @@ package org.collector.presentation.dto;
 
 import java.time.LocalDateTime;
 
-import org.collector.domain.CycleInfo;
-import org.collector.domain.VehicleInformation;
 import org.hibernate.validator.constraints.Range;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
@@ -42,17 +40,4 @@ public record CListRequest(
 	@Range(min = 0, max = 9999, message = "배터리 전압은 0 ~ 9999 사이여야 합니다.")
 	Integer bat
 ) {
-	public static CycleInfo from(CListRequest request, VehicleInformation vehicleInformation) {
-		return CycleInfo.builder()
-			.interval_at(request.interval_at())
-			.gcd(request.gcd())
-			.lat(CycleInfo.convertToSixDecimalPlaces(request.lat()))
-			.lon(CycleInfo.convertToSixDecimalPlaces(request.lon()))
-			.ang(request.ang())
-			.spd(request.spd())
-			.sum(request.sum())
-			.bat(request.bat())
-			.vehicleInformation(vehicleInformation)
-			.build();
-	}
 }
