+++ monicar-collector/src/main/java/org/collector/presentation/dto/CycleInfoRequest.java
@@ -3,7 +3,6 @@ package org.collector.presentation.dto;
 import java.util.List;
 
 import org.collector.common.annotation.MatchCycleSize;
-import org.collector.domain.VehicleInformation;
 import org.hibernate.validator.constraints.Range;
 
 import jakarta.validation.Valid;
@@ -34,13 +33,4 @@ public record CycleInfoRequest(
 	@Valid
 	List<CListRequest> cList
 ) {
-	public static VehicleInformation from(CycleInfoRequest request) {
-		return VehicleInformation.builder()
-			.mdn(request.mdn())
-			.tid(request.tid())
-			.mid(request.mid())
-			.pv(request.pv())
-			.did(request.did())
-			.build();
-	}
 }
